requirements.txt
# Core ML
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.10.0+cpu
torchaudio==2.10.0+cpu

# HF stack
transformers==4.57.6
peft==0.18.1
accelerate==1.12.0
huggingface-hub==0.34.0
safetensors==0.7.0
tokenizers==0.22.2

# Data / metrics
numpy==2.2.6
soundfile==0.13.1
jiwer==4.0.0

# Utilities
PyYAML==6.0.3
requests==2.32.5
tqdm==4.67.3
packaging==26.0
